Give up and add -Wno-unused-result to the Travis environment.
The problem seems to be that Travis forces _FORTIFY_SOURCE, which enables __warn_unused_result__ annotations in glibc.  Some of those annotations are of dubious value; in the case of asprintf(3) and vasprintf(3), they flag code that doesn't check the return value as unsafe even if it checks the pointer instead (which is guaranteed to be NULL in case of failure, and arguably more useful than the return value).

Unfortunately, gcc explicitly and intentionally ignores (void) casts:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66425

so we have no choice but to quench the warning.